Separate out viewer code and cleanup.
